# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for managing Cloud Hypervisor VMs.",
        "title": "Cloud Hypervisor API",
        "termsOfService": "https://github.com/org/ch-api",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/org/ch-api/issues",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/vms/{id}/interfaces": {
            "post": {
                "description": "Attaches a new network interface to the virtual machine.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Network"
                ],
                "summary": "Add Interface",
                "parameters": [
                    {
                        "type": "string",
                        "description": "VM ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Interface configuration",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.AddInterfaceRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.InterfaceResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    },
                    "404": {
                        "description": "VM not found",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    }
                }
            }
        },
        "/api/v1/vms/{id}/interfaces/{iface_id}": {
            "delete": {
                "description": "Detaches the network interface from the virtual machine.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Network"
                ],
                "summary": "Remove Interface",
                "parameters": [
                    {
                        "type": "string",
                        "description": "VM ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Interface ID",
                        "name": "iface_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.InterfaceResponse"
                            }
                        }
                    },
                    "404": {
                        "description": "VM or interface not found",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates fields of an existing network interface on the virtual machine.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Network"
                ],
                "summary": "Patch Interface",
                "parameters": [
                    {
                        "type": "string",
                        "description": "VM ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Interface ID",
                        "name": "iface_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Interface update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.PatchInterfaceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.InterfaceResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    },
                    "404": {
                        "description": "VM or interface not found",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.AddInterfaceRequest": {
            "type": "object",
            "properties": {
                "ip": {
                    "type": "string"
                },
                "mac": {
                    "type": "string"
                },
                "mask": {
                    "type": "string"
                },
                "tap": {
                    "type": "string"
                }
            }
        },
...
        "handler.InterfaceResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "mac": {
                    "type": "string"
                },
                "mask": {
                    "type": "string"
                },
                "tap": {
                    "type": "string"
                }
            }
        },
...
        "handler.PatchInterfaceRequest": {
            "type": "object",
            "properties": {
                "ip": {
                    "type": "string"
                },
                "mac": {
                    "type": "string"
                },
                "mask": {
                    "type": "string"
                },
                "tap": {
                    "type": "string"
                }
            }
        },
...
        "problem.Detail": {
            "type": "object",
            "properties": {
                "detail": {
                    "description": "Detail is a human-readable explanation specific to this occurrence.",
                    "type": "string"
                },
                "errors": {
                    "description": "Errors holds machine-readable validation errors.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/problem.FieldError"
                    }
                },
                "instance": {
                    "description": "Instance is a URI reference that identifies the specific occurrence.",
                    "type": "string"
                },
                "status": {
                    "description": "Status is the HTTP status code generated by the origin server.",
                    "type": "integer"
                },
                "title": {
                    "description": "Title is a short, human-readable summary of the problem type.",
                    "type": "string"
                },
                "type": {
                    "description": "Type is a URI reference that identifies the problem type.",
                    "type": "string"
                }
            }
        },
        "problem.FieldError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
...
}